always trigger apply-sysctl in case failure in common role
sack=1 is configured in common role(networking.yml)
If task in common role fails, it may case that sack=1 is configured in
sysctl config file, while it's not applied in OS. Because it can't
trigger the apply sysctl handler.
Re-run will not be able to trigger apply sysctl, because sysctl
configure files will not change in re-run, then sack=1 will never be
applied until reboot or do sysctl -p manually.
 
As too many tasks are in common role, the failure probability of common
role is relative high.