Support adding a failed disk back to Hybrid Ceph Cluster
We have several osds on each osd node.
For hybrid ceph, we can't run ursula deploy against a single osd.
We have to run against all osds on the node.

So we can't add a failed disk back to hybrid ceph cluster by ursula.

This patch is to support this feature.