make chown be faster
we change owner of files in /var/lib/ceph in ceph-update roles, this is
very time consuming when ceph has large mount of data in it.
the bottleneck is the IO, there are so many random read requests when
chown.
This patch is to add a new playbook for pre-reading the metadata of
files. So that the chown will be much faster.